+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  
+- Add first biomes and their settings
 - Use of bevy fast tilemap
 
 ## [0.1.0] - 2024-10-22